Remove copts
Upon reconsideration of my options, I think that "common options" that
apply globally to all of uwp will either be too rare or too specialised
to the function of the specific program to actually apply. My test case
was a flag for the site, but this raised the question of which tools
don't actually require a site to operate on, and what will those that do
perform if no such flag is provided? Flags are supposed to be optional.

As a result, I have simplified and deleted over 45 minutes of unstaged
work (don't make me regret it now). Options will now be solely handled
by application specific code.